requirements.txt
moviepy==1.0.3
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
scipy==1.11.4
Pillow==10.1.0
numpy==1.24.3
